+++ Src/Cognate/Models/TestVariantScore.cs
@@ -5,7 +5,7 @@ namespace Cognate.Models
 	[TableName("cognateTestVariantScore")]
 	internal class TestVariantScore : Entity
 	{
-		public int TestId { get; set; }
+		public long TestId { get; set; }
 		public int VariantId { get; set; }
 		public int Score { get; set; }
 	}
